+++ src/core/drd_config.h
@@ -42,7 +42,6 @@ const gchar *drd_config_get_private_key_path(DrdConfig *self);
 const gchar *drd_config_get_nla_username(DrdConfig *self);
 const gchar *drd_config_get_nla_password(DrdConfig *self);
 gboolean drd_config_is_nla_enabled(DrdConfig *self);
-gboolean drd_config_get_system_mode(DrdConfig *self);
 DrdRuntimeMode drd_config_get_runtime_mode(DrdConfig *self);
 const gchar *drd_config_get_pam_service(DrdConfig *self);
 guint drd_config_get_capture_width(DrdConfig *self);